Refactor: encapsule Article component from article page
